challenge.go: move mock stuff into challenge_test.go
